Final commit code
+ Bug fixes and small code refactorings
+ Last console view implementations

+++ Code/src/main/java/com/github/Elmicass/SFJTeam_Casotto/services/EquipmentServices.java
@@ -33,20 +33,27 @@ public class EquipmentServices implements IEquipmentServices {
     }
 
     @Override
+    public Equipment save(Equipment eq) {
+        return eqRepository.save(eq);
+    }
+
+    @Override
     public boolean createEquipment(@NonNull String name, @NonNull String description, @NonNull String type) {
         if (equipmentCreationErrorsCheckin(type)) {
             Equipment equipment =  new Equipment(name, description, type);
-            eqRepository.save(equipment);
+            save(equipment);
             return true;
         } else return false;
     }
     
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Integer id) throws IllegalStateException, EntityNotFoundException, IllegalArgumentException {
         if (id.toString().isBlank())
             throw new IllegalArgumentException("The equipment ID is empty");
         if (!(exists(id)))
             throw new EntityNotFoundException("The equipment with ID: " + id + " does not exist");
+        if (!(getInstance(id).getActivity().isEmpty()))
+            throw new IllegalStateException("It is not possible to delete an equipment when there are activities scheduled to use it!");
         eqRepository.deleteById(id);
         return !exists(id);
     }
